02-use-cases/data_analysis_with_code/agentkit_deploy/requirements.txt
veadk-python
veadk-python[extensions]
google-adk
python-dotenv
lancedb
agentkit-sdk-python
volcengine-python-sdk[ark]
openai
pyarrow
httpx[http2]
httpx
duckdb
